third_party/blink/web_tests/platform/linux/virtual/strict-ipv4-embedded-ipv6/external/wpt/url/url-constructor.any_include=mailto-expected.txt

This is a testharness.js-based test.
PASS Loading data…
PASS Parsing: <mailto:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <mailto:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <mailto:/example.com/> against <about:blank>
PASS Parsing: <mailto:example.com/> against <about:blank>
PASS Parsing: <#x> against <mailto:x@x.com>
FAIL Parsing: <mailto:/../> against <about:blank> assert_equals: href expected "mailto:/" but got "mailto:/../"
PASS Parsing: <> against <about:blank>
FAIL Parsing: <mailto://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: href expected "mailto://example.com:8080/pathname?search#hash" but got "mailto://example.com:8080/pathname?search%23hash"
FAIL Parsing: <mailto:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <mailto://test/a/../b> against <about:blank> assert_equals: href expected "mailto://test/b" but got "mailto://test/a/../b"
FAIL Parsing: <mailto://:443> against <about:blank> assert_throws_js: function "function() {
          bURL(expected.input, expected.base)
        }" did not throw
FAIL Parsing: <mailto://test:test> against <about:blank> assert_throws_js: function "function() {
          bURL(expected.input, expected.base)
        }" did not throw
FAIL Parsing: <mailto://[:1]> against <about:blank> assert_throws_js: function "function() {
          bURL(expected.input, expected.base)
        }" did not throw
Harness: the test ran to completion.
